requirements2.txt
django==2.2.4
scikit-allel==1.3.3
lazypredict==0.2.5
pytest-docker==0.10.1
celery==4.2.0
django-celery>3.1.15, <4.0
django-celery-results==1.0.1
djangorestframework
vine==1.3.0
kombu==4.2.0
redis
